Allowing autoscaling
Added window height and window width parameters.
If these are set and scale is set to 0 then it will autoscale (based on the height) to fit
If they are set and scale is NOT set to 0 then it will draw in the middle of the screen, possibly leaving a border.

Also: setScale has been removed because it's a headache
Updated main.lua to respect changes
Updated README.md